requirements.txt
streamlit==1.33.0
requests
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.0.1
huggingface-hub==0.16.4
supabase
python-dotenv
numpy
scikit-learn
scipy